[backend] remove file-storage-helper
