+++ opencti-platform/opencti-graphql/src/database/middleware.js
@@ -224,7 +224,7 @@ import {
 } from '../utils/confidence-level';
 import { buildEntityData, buildInnerRelation, buildRelationData } from './data-builder';
 import { isIndividualAssociatedToUser, verifyCanDeleteIndividual, verifyCanDeleteOrganization } from './data-consistency';
-import { deleteAllObjectFiles, moveAllFilesFromEntityToAnother, uploadToStorage } from './file-storage-helper';
+import { deleteAllObjectFiles, moveAllFilesFromEntityToAnother, uploadToStorage } from './file-storage';
 import { getFileContent } from './raw-file-storage';
 import { storeFileConverter } from './file-storage';
 import { getDraftContext } from '../utils/draftContext';
