+++ opencti-platform/opencti-graphql/tests/02-integration/01-database/index-file-test.js
@@ -7,7 +7,7 @@ import { ADMIN_USER, testContext } from '../../utils/testQuery';
 import { getFileContent } from '../../../src/database/raw-file-storage';
 import { INDEX_FILES } from '../../../src/database/utils';
 import { resetFileIndexing } from '../../../src/domain/indexedFile';
-import { uploadToStorage } from '../../../src/database/file-storage-helper';
+import { uploadToStorage } from '../../../src/database/file-storage';
 import { getManagerConfigurationFromCache, updateManagerConfigurationLastRun } from '../../../src/modules/managerConfiguration/managerConfiguration-domain';
 import { SYSTEM_USER } from '../../../src/utils/access';
