+++ opencti-platform/opencti-graphql/tests/02-integration/02-resolvers/draftWorkspace-test.ts
@@ -5,7 +5,7 @@ import { ADMIN_USER, adminQuery, queryAsAdmin, testContext } from '../../utils/t
 import { MARKING_TLP_GREEN, MARKING_TLP_RED } from '../../../src/schema/identifier';
 import { buildDraftValidationBundle } from '../../../src/modules/draftWorkspace/draftWorkspace-domain';
 import { DRAFT_VALIDATION_CONNECTOR_ID } from '../../../src/modules/draftWorkspace/draftWorkspace-connector';
-import { fileToReadStream } from '../../../src/database/file-storage-helper';
+import { fileToReadStream } from '../../../src/database/file-storage';
 import { STIX_EXT_OCTI } from '../../../src/types/stix-2-1-extensions';
 import { DRAFT_STATUS_OPEN, DRAFT_STATUS_VALIDATED } from '../../../src/modules/draftWorkspace/draftStatuses';
