+++ opencti-platform/opencti-graphql/tests/utils/globalSetup.ts
@@ -17,7 +17,7 @@ import { initializeInternalQueues } from '../../src/database/rabbitmq';
 import { executionContext } from '../../src/utils/access';
 import { initializeData } from '../../src/database/data-initialization';
 import { shutdownModules, startModules } from '../../src/managers';
-import { deleteAllBucketContent } from '../../src/database/file-storage-helper';
+import { deleteAllBucketContent } from '../../src/database/file-storage';
 import { initExclusionListCache } from '../../src/database/exclusionListCache';
 import { initLockFork } from '../../src/lock/master-lock';
