+++ opencti-platform/opencti-graphql/tests/utils/testQueryHelper.ts
@@ -20,7 +20,7 @@ import { streamConverter } from '../../src/database/file-storage';
 import conf, { logApp } from '../../src/config/conf';
 import { AUTH_REQUIRED, FORBIDDEN_ACCESS } from '../../src/config/errors';
 import { getSettings, settingsEditField } from '../../src/domain/settings';
-import { fileToReadStream } from '../../src/database/file-storage-helper';
+import { fileToReadStream } from '../../src/database/file-storage';
 import { resetCacheForEntity } from '../../src/database/cache';
 import { ENTITY_TYPE_SETTINGS } from '../../src/schema/internalObject';
